refactor: вынесла енамы в отдельный пакет, т.к. их с каждым заданием становится всё больше + схлопнула два if в один.
