+++ src/main/java/com/tcs/edu/enums/Doubling.java
@@ -1,0 +1,5 @@
+package com.tcs.edu.enums;
+
+public enum Doubling {
+    DOUBLES, DISTINCT
+}
